+++ Neocom/NCDatabaseGroupPickerViewContoller.h
@@ -9,7 +9,7 @@
 #import "NCTableViewController.h"
 
 @interface NCDatabaseGroupPickerViewContoller : NCTableViewController
-@property (nonatomic, assign) NSInteger categoryID;
+@property (nonatomic, assign) int32_t categoryID;
 @property (nonatomic, strong) EVEDBInvGroup* selectedGroup;
 
 @end
